+++ ft_strncmp.c
@@ -6,12 +6,13 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 15:16:03 by ppontet           #+#    #+#             */
-/*   Updated: 2024/11/06 15:03:03 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2024/11/07 16:24:31 by ppontet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
+//@TODO RATE UN TEST
 /**
  * @brief Find differences between first and second
  *
@@ -23,7 +24,7 @@
 
 int	ft_strncmp(const char *first, const char *second, size_t length)
 {
-	unsigned int	index;
+	size_t	index;
 
 	index = 0;
 	if (length == 0)
